// frontend/src/app/models/user.model.ts
export interface User {
  userId: string
  username: string
  password: string
  importanceGenre: number
  importanceActor: number
  importanceDirector: number
  importanceRating: number
  importanceLanguage: number
  importanceCountry: number
  importanceYear: number
  importanceDuration: number
}